added AudioOverlay > Server Selection
The server selection dropdown should probably be extracted out into a
component but for now this feels fine. I didn't know you could so easily
build a dropdown component in Angular.

I realise that my issue with EMPTY_PLAYLIST, EMPTY_SONG, etc. was that I
needed a Subject, not a BehaviourSubject.