added AudioOverlay > Disconnect Audio
This implementation does introduce a bug that I need to fix.
When you create ServerAudio in one Server, then switch to another
Server, the Disconnect button will still be present.

Additionally, when starting up the App, if there was already Server
audio present in the Server the button will not appear.

This is all because we are trying to manage state on the frontend when
we should be querying the backend for the actually state. The correct
flow would be: Whenever a Server is selected, find out whether it has
ServerAudio.