added AudioOverlay > ServiceStatus
This component will eventually become the ServiceSelector, though I am
not sure if that will be part of this iteration of the design - I still
feel it can be.

Now that we fetch the ServiceStatus, we should be able to revise the
base AppComponent and use ngTemplates to change the layout from
catalogue building only to one where the Catalogue is nested within the
AudioOverlay.