extracted ServerAudioStatus
This implementation is a workaround at best. Until the AudioService API
is truly asynchronous, we cannot rely on being able to send a request to
fetch the new state of the ServerAudio after we POST or PATCH.

In fact, I need to make sure we PATCH if we know we already have a
ServerAudio.

Another way we could do this is to use websockets and simply listen for
pushes when the ServerAudio status changes. However, websockets are not
part of this iteration of the design.