mv tile selectors -> `article` attribute-selectors
This change is meant to set a new standard of using ng attribute selectors
when defining my UI components.

Tile components, and most components I've written, are wrapped in a custom
app***component HTML tag that nests the actual HTML element the component is
meant to be representing - I'd prefer if only the Controller/Smart
components had their own custom tag (as far as possible).