AudioOverlay>Aside: mv grid->flex
I just like grid layouts, but there are times where its not the right
tool - this is one of those times. The aside menu has dynamically
rendering components, and components that need to resize.

I wanted to use the grid because I needed three elements to have 'fixed'
positions on the aside, but that can be accomplished easier (and better)
with a flex layout and margin:auto.