[fixes] default font undefined
Just using Tahoma, sans-serif for now until I can figure out the font
choices for the application. This is needed for now, atleast, as I
noticed the fonts in different components are not consistent. I would
have thought that the browser has a single default, but it seems to pick
a different font depending on the html around it?