added visual indicator for selected Channel
Mimics the Discord design and displays a row with the AudioService as a
'user' in the Channel.

This does duplicate the information available on the AudioOverlay,
but it feels natural to see it there as the Aside is meant to mimic the
Discord UI which does the same.

This will also be useful behaviour for indicating whether there are any
other ServerAudios connected to the Channel (when we eventually support
both the 'ambience' and 'background music' AudioServices.