[fix] DialogForm autofocusing
The HTML autofocus attribute only works once on page load, which is a problem in an SPA like Angular.

This is a workaround to apply focus to an HTMLElement when the Dialog is opened.

NOTE: A Directive would be ideal, but focusing requires the component to be rendered which is a problem
for all my current inputs that are hidden inside a Dialog.