added TextSearchInputDirective
As there will be search boxes on almost all the screens, it makes sense
to create a utility for the basic needs of a text search input field.
Rather than try create a component, I feel a Directive that 'enchances'
the native input element with the functionality I need is preferable. I
can let other Components style and configure the input as needed.