fix @angular-eslint/prefer-inject
Migrated to inject() usage. Had to do this one manually, as the
migration was not in the collections.json and I didn't want to install
the schematic tool just to run the inject-migration. The pattern I was
using was from way back in ng14 I believe and this project was started
in ng18 so I guess that migration was no longer needed in
collections.json.

Explict public constructors are needed for components on the routes
because of strict Typescript type requiring a type with a public
constructor.