added linting & formatting into Makefile build recipe
This is good enough for now but as the project gets bigger we can use
things like, `git diff --name-only HEAD`, etc. to apply the formatting
and linting to only what has changed.

Probably worth mentioning that the Makefile and pre-commit script aren't
perfect and won't enforce the workflow absolutely. They are just good
automations for the flow and will help me catch if I am slipping up.