added AudioOverlayToggleComponent
The ability to toggle the AudioOverlay is essential to the whole idea of
this application. Up until now, I have been 'toggling' it off by simply
turning off the actual service. The 'jumping' off the layout due to the
toggle can only be resolved through (transition) animations that I am
not willing to experiment with at the moment. For this pass, just having
the ability to toggle the Overlay via the UI will suffice.

[revisions]
- AppComponent now defaults to a layout without an AudioOverlay. Editing
  the Catalogue is the primary interaction on the App, with the Audio
  interactions occuring ocassionally on game nights, etc.
- allow HeaderComponent to accept content. Styling changed to a flex
  layout with the Heading always pushed left.

[additions]
- extracted DebounceDirective from TextSearchInputDirective.

[housekeeping]
- mv TextSearchInputDirective InputSearchDebounceDirective.