mv FormDialogComponets FormComponents
This continues the theme of trying to use better HTML semantics.
I realised the DOM for the previous "FormDialogs" didn't reflect what they were.
Also, the FormDialogComponents were out-of-place with the rest of
the simpler UI components in the app.

This revision should better reflect what we are trying to model in code;
There are several Dialogs that are possible with a User on a Page.
The User is using the Catalogue to make requests to Create, or Update resources.
These requests are captured on Forms we submit on their behalf.

[revisions]
- added #focus, #submit to Form interface. See interface and usages.
- extracted .modal style class. It can be applied to a `form` as well,
  which is what I've done. The form is inside an `app-dialog`, it is a
  "modal" form. I feel this makes more sense.
- mv `app-overlay` to first-child  in `app-dialog` DOM node. This makes
  more sense to think about the overlay appearing first and masking
  everything else.
- mv `app-dialog`(s) to end of `app-*-page` DOMs. The dialogs are
  occuring on the pages themselves. They should not be nested in other
  containers.

[removals]
- rm unneccessary `div` in the OverlayComponent. It seems this was
  before I discovered `:host` styling.