fix AppHeaderComponent height
Because the component went from `display: block` to `display: flex`,
and was inside the `app-layout` class which was a flex container,
it became a flex child that could shrink. This meant it did not consume
the full height specified by the `.header-bar` class.

The fix was to define the `flex-shrink` rule that will apply if the
`.header-bar` element becomes a flex-child.

[revisions]
- rm host binding directive in HeaderComponent. In this case we can
  easily apply the style traditionally.