maintain Catalogue search filter between page changes (#8)
maintain catalogue search filter between page changes

As a User, when I put a term in the search bar and then navigate into a
Playlist, I actually expect the search to remain because I use the
search more like a filter on what Playlists I am looking at in the
Catalogue.

[fixes]
- on Playlist creation, the search filter was being ignored when
Playlists were refetched. As a result, the current search filter became
unusable. This occurred because the TextInputSearchDirective was
designed not to emit output unless the input changes from its initial
value. The fix ensures that, when we refetch Playlists after creation,
the current search filter is reused—if present.

[revisions]
- TextInputSearchDirective now captures the initial value of the input
field. Previously, we assumed the initial state to be empty. This
revision is required because we must now set the initial value of the
search filter back to the state it previously was when navigating back
to the CataloguePlaylistsPage.
- added refresh semantics for PageComponents. The data we load for a
page should be reactive to direct user interaction (e.g., applying a
filter), but we need to provide a way for indirect interaction to
trigger loading (e.g., creating a new item, which means we should
refetch the items).

[additions]
- added CatalogueStateService to manage search state. The state cannot
be tied to the page or search component themselves, because they are
destroyed on navigation off the page.

[housekeeping]
- section separators on CataloguePlaylistsPage. Still need to give more
thought to how we should group fields and methods, and other general
conventions.