makefiles-v1.3.^2;angular-v^2.0.0
[patch] makefiles. Fixes "make image" not remaking on build change.

Docker targets should be defined after then <technology> section, because the
building of the development images is dependant on the artifacts declared by
the <technology> section.

[revision] makefiles-angular. Serve from an nginx container.

"ng serve", even with all the flags we were using, still wasn't what we
wanted. Even after we ran a "make build", "ng serve" would not serve
those dist/ files, and instead we'd need to "ng serve" again and it
would rebuild.

If our flow already creates the build artifacts, the simplest way to
serve them is via an actual static file server. We just mount the
output directory, and it automatically will serve updated files after a
"make build".

- [removals] rm stop-process.sh. No longer needed due to serving via
  nginx, not an "ng serve" process.