added AudioStateService (#11)
A single-source of truth for the state of the Discord Bot. Discord
components will now interface directly with the StateService, rather
than wiring inputs and outputs through several layers of components.

[revisions]
- features/audio components now interface with the AudioStateService,
rather than receiving inputs and outputting events to be handled in
AudioServiceOverlay.
- AudioApiService#getAudioService now transforms all failures into
"undefined". If the Discord service is down, we want to return
"undefined" and not fail, as this is an expected circumstance.

[housekeeping]
- correct naming, layout in AudioApiService.