reverse proxy via Nginx
By allowing the UI to send its XHR requests to proxies "discord" and
"storage", rather than directly to the APIs, we can access the UI from
another device on the local network.

Previously, the UI would call out to "localhost", which would fail when
called from a device that was not the host.

[removals]
- rm "environment.development". The multiple environment setup was
  only for specifying different hosts/API urls. This is now
  superseded by reverse proxying.

[additions]
- added ".env" target to "make project". This is the best way I can find
  to keep configuration synced to the remote, without actually
  hardcoding the IP of the workmachine and commiting it to the
  repository. Manual configuration can't be avoided, but some of it can
  be automated.