use HTML dialogs (#14)
The HTML Dialog element natively supports almost all functionality we
manually defined with DialogComponent and OverlayComponent. It has
native integrations with HTML Form, automatically focusing input fields
when opened and closing the dialog when the form is submitted.

[additions]
- added ModalDirective. For mobile compatibility, dismiss the modal if
the user taps outside. This was the only functionality not already
provided by HTML Dialog.